refactor: (#256) dto로 변경
