+++ dms-application/src/main/kotlin/team/aliens/dms/domain/point/dto/PointHistoryDto.kt
@@ -1,9 +1,9 @@
-package team.aliens.dms.domain.point.spi.vo
+package team.aliens.dms.domain.point.dto
 
 import team.aliens.dms.domain.point.model.PointType
 import java.time.LocalDate
 
-data class PointHistoryVO(
+data class PointHistoryDto(
     val date: LocalDate,
     val type: PointType,
     val name: String,
