+++ dms-application/src/main/kotlin/team/aliens/dms/domain/point/dto/QueryPointHistoryResponse.kt
@@ -1,8 +1,6 @@
 package team.aliens.dms.domain.point.dto
 
-import team.aliens.dms.domain.point.spi.vo.PointHistoryVO
-
 data class QueryPointHistoryResponse(
     val totalPoint: Int,
-    val points: List<PointHistoryVO>
+    val points: List<PointHistoryDto>
 )
