+++ dms-application/src/main/kotlin/team/aliens/dms/domain/point/dto/QueryStudentPointHistoryResponse.kt
@@ -1,7 +1,5 @@
 package team.aliens.dms.domain.point.dto
 
-import team.aliens.dms.domain.point.spi.vo.PointHistoryVO
-
 data class QueryStudentPointHistoryResponse(
     val pointHistories: List<PointHistoryVO>
 )
