+++ dms-infrastructure/src/main/kotlin/team/aliens/dms/persistence/point/PointHistoryPersistenceAdapter.kt
@@ -4,11 +4,11 @@ import com.querydsl.jpa.impl.JPAQueryFactory
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Component
 import team.aliens.dms.common.dto.PageData
+import team.aliens.dms.domain.point.dto.PointHistoryDto
 import team.aliens.dms.domain.point.dto.QueryAllPointHistoryResponse
 import team.aliens.dms.domain.point.model.PointHistory
 import team.aliens.dms.domain.point.model.PointType
 import team.aliens.dms.domain.point.spi.PointHistoryPort
-import team.aliens.dms.domain.point.spi.vo.PointHistoryVO
 import team.aliens.dms.persistence.point.entity.QPointHistoryJpaEntity.pointHistoryJpaEntity
 import team.aliens.dms.persistence.point.mapper.PointHistoryMapper
 import team.aliens.dms.persistence.point.repository.PointHistoryJpaRepository
@@ -65,7 +65,7 @@ class PointHistoryPersistenceAdapter(
         type: PointType?,
         isCancel: Boolean?,
         pageData: PageData
-    ): List<PointHistoryVO> {
+    ): List<PointHistoryDto> {
         return queryFactory
             .select(
                 QQueryPointHistoryVO(
@@ -87,7 +87,7 @@ class PointHistoryPersistenceAdapter(
             .orderBy(pointHistoryJpaEntity.createdAt.desc())
             .fetch()
             .map {
-                PointHistoryVO(
+                PointHistoryDto(
                     date = it.date.toLocalDate(),
                     type = it.pointType,
                     name = it.pointName,
